Tests: shut up warnings for TODO tests.
Recent versions of Test::Simple/More/Builder (0.61+) throw warnings on undef
in like().  While this is good for working tests, it's not for TODO ones.

Other tricky part is that neither "no warnings" pragma nor $^W can be used to
avoid warnings.  The only way that works is to redefine $SIG{__WARN__}.